testclasses with more than one class in one file splitted in more files
