+++ test/Adapter/JsPushStub.php
@@ -1,0 +1,27 @@
+<?php
+/**
+ * Zend Framework (http://framework.zend.com/)
+ *
+ * @link      https://github.com/zendframework/zendframework for the canonical source repository
+ * @copyright Copyright (c) 2005-2017 Zend Technologies USA Inc. (http://www.zend.com)
+ * @license   http://framework.zend.com/license/new-bsd New BSD License
+ */
+
+namespace ZendTest\ProgressBar\Adapter;
+
+use Zend\ProgressBar\Adapter;
+
+class JsPushStub extends Adapter\JsPush
+{
+    protected $_lastOutput = null;
+
+    public function getLastOutput()
+    {
+        return $this->_lastOutput;
+    }
+
+    protected function _outputData($data)
+    {
+        $this->_lastOutput = $data;
+    }
+}
