+++ test/MockUp.php
@@ -1,0 +1,66 @@
+<?php
+/**
+ * Zend Framework (http://framework.zend.com/)
+ *
+ * @link      https://github.com/zendframework/zendframework for the canonical source repository
+ * @copyright Copyright (c) 2005-2017 Zend Technologies USA Inc. (http://www.zend.com)
+ * @license   http://framework.zend.com/license/new-bsd New BSD License
+ */
+
+namespace ZendTest\ProgressBar;
+
+use Zend\ProgressBar\Adapter;
+
+class MockUp extends Adapter\AbstractAdapter
+{
+    protected $_current;
+    protected $_max;
+    protected $_timeTaken;
+    protected $_timeRemaining;
+    protected $_text;
+    protected $_percent;
+
+    public function notify($current, $max, $percent, $timeTaken, $timeRemaining, $text)
+    {
+        $this->_current       = $current;
+        $this->_max           = $max;
+        $this->_percent       = $percent;
+        $this->_timeTaken     = $timeTaken;
+        $this->_timeRemaining = $timeRemaining;
+        $this->_text          = $text;
+    }
+
+    public function finish()
+    {
+    }
+
+    public function getCurrent()
+    {
+        return $this->_current;
+    }
+
+    public function getMax()
+    {
+        return $this->_max;
+    }
+
+    public function getPercent()
+    {
+        return $this->_percent;
+    }
+
+    public function getTimeTaken()
+    {
+        return $this->_timeTaken;
+    }
+
+    public function getTimeRemaining()
+    {
+        return $this->_timeRemaining;
+    }
+
+    public function getText()
+    {
+        return $this->_text;
+    }
+}
