+++ test/ProgressBarTest.php
@@ -93,95 +93,3 @@ class ProgressBarTest extends TestCase
         return new Stub(new MockUp(), $min, $max, $persistenceNamespace);
     }
 }
-
-class Stub extends \Zend\ProgressBar\ProgressBar
-{
-    public function sleep($seconds)
-    {
-        $this->startTime -= $seconds;
-    }
-
-    public function getCurrent()
-    {
-        return $this->adapter->getCurrent();
-    }
-
-    public function getMax()
-    {
-        return $this->adapter->getMax();
-    }
-
-    public function getPercent()
-    {
-        return $this->adapter->getPercent();
-    }
-
-    public function getTimeTaken()
-    {
-        return $this->adapter->getTimeTaken();
-    }
-
-    public function getTimeRemaining()
-    {
-        return $this->adapter->getTimeRemaining();
-    }
-
-    public function getText()
-    {
-        return $this->adapter->getText();
-    }
-}
-
-class MockUp extends \Zend\ProgressBar\Adapter\AbstractAdapter
-{
-    protected $_current;
-    protected $_max;
-    protected $_percent;
-    protected $_timeTaken;
-    protected $_timeRemaining;
-    protected $_text;
-
-    public function notify($current, $max, $percent, $timeTaken, $timeRemaining, $text)
-    {
-        $this->_current       = $current;
-        $this->_max           = $max;
-        $this->_percent       = $percent;
-        $this->_timeTaken     = $timeTaken;
-        $this->_timeRemaining = $timeRemaining;
-        $this->_text          = $text;
-    }
-
-    public function finish()
-    {
-    }
-
-    public function getCurrent()
-    {
-        return $this->_current;
-    }
-
-    public function getMax()
-    {
-        return $this->_max;
-    }
-
-    public function getPercent()
-    {
-        return $this->_percent;
-    }
-
-    public function getTimeTaken()
-    {
-        return $this->_timeTaken;
-    }
-
-    public function getTimeRemaining()
-    {
-        return $this->_timeRemaining;
-    }
-
-    public function getText()
-    {
-        return $this->_text;
-    }
-}
